2007-03-02 03:30 UTC+0100 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/source/rdd/dbf1.c   * harbour/source/rdd/delim1.c   * harbour/source/rdd/sdf1.c   * harbour/source/rdd/dbfcdx/dbfcdx1.c   * harbour/source/rdd/dbffpt/dbffpt1.c   * harbour/source/rdd/hbsix/sxcompr.c   * harbour/source/rdd/hsx/hsx.c   * harbour/source/rdd/usrrdd/usrrdd.c   * harbour/source/rtl/cdpapi.c     * casting

+++ harbour/source/rdd/dbf1.c
@@ -1499,7 +1499,7 @@ static ERRCODE hb_dbfGetValue( DBFAREAP pArea, USHORT uiIndex, PHB_ITEM pItem )
                   iLen = 10;
                   break;
                case 8:
-                  dValue = HB_GET_LE_INT64( pArea->pRecord + pArea->pFieldOffset[ uiIndex ] );
+                  dValue = ( double ) HB_GET_LE_INT64( pArea->pRecord + pArea->pFieldOffset[ uiIndex ] );
                   iLen = 20;
                   break;
                default:
